requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-dotenv>=1.0.0
sqlmodel>=0.0.8
google-generativeai>=0.8.5
pyttsx3>=2.90
twilio>=8.8.0
python-multipart>=0.0.6
aiofiles>=23.1.0